+++ girder_annotation/girder_large_image_annotation/utils/__init__.py
@@ -29,6 +29,7 @@ dataFileExtReaders = {
     'application/x-xls': 'read_excel',
 }
 scanDatafileRecords = 50
+scanAnnotationElements = 5000
 
 
 @functools.lru_cache(maxsize=100)
@@ -393,7 +394,8 @@ class PlottableItemData:
     maxDistinct = 20
     allowedTypes = (str, bool, int, float)
 
-    def __init__(self, user, item, annotations=None, adjacentItems=False, sources=None):
+    def __init__(self, user, item, annotations=None, adjacentItems=False,
+                 sources=None, compute=None):
         """
         Get plottable data associated with an item.
 
@@ -408,15 +410,32 @@ class PlottableItemData:
         :param sources: None for all, or a string with a comma-separated list
             or a list of strings; when a list, the options are folder, item,
             annotation, datafile.
+        :param compute: None for none, or a dictionary with keys "columns": a
+            list of columns to include in the computation; if unspecified or an
+            empty list, no computation is done, "function": a string with the
+            name of the function, such as umap, "params": additional parameters
+            to pass to the function.  If none of the requiredKeys are
+            compute.(x|y|z), the computation will not be performed.  Only rows
+            which have all selected columns present will be included in the
+            computation.
         """
         self.user = user
         self._columns = None
         self._datacolumns = None
         self._data = None
+        self._compute = None
+        try:
+            if len(compute['columns']):
+                self._compute = {'function': 'umap', 'params': {
+                    'random_state': 1, 'n_jobs': 1}}
+                self._compute.update(compute)
+        except Exception:
+            pass
         if sources and not isinstance(sources, (list, tuple)):
             sources = sources.split(',')
         self._sources = tuple(sources) if sources else None
-        if self._sources and 'annotation' not in self._sources:
+        if (self._sources and 'annotation' not in self._sources and
+                'annotationelement' not in self._sources):
             annotations = None
         self._fullScan = adjacentItems == '__all__'
         self._findItems(item, adjacentItems)
@@ -559,7 +578,11 @@ class PlottableItemData:
         'bbox.y0': 'Bounding Box Low Y',
         'bbox.x1': 'Bounding Box High X',
         'bbox.y1': 'Bounding Box High Y',
+        'compute.x': 'Dimension Reduction X',
+        'compute.y': 'Dimension Reduction Y',
+        'compute.z': 'Dimension Reduction Z',
     }
+    computeColumns = {'compute.x', 'compute.y', 'compute.z'}
 
     def itemNameIDSelector(self, isName, selector):
         """
@@ -602,14 +625,14 @@ class PlottableItemData:
         def annotationElementSelector(record, data, row):
             bbox = [col[1](record, data, row) for col in cols]
             if key in self._datacolumns:
-                for row in self._datacolumns[key]:
-                    if self._datacolumns[key][row] is not None:
+                for srow in self._datacolumns[key]:
+                    if self._datacolumns[key][srow] is not None:
                         for bidx, bkey in enumerate(['bbox.x0', 'bbox.y0', 'bbox.x1', 'bbox.y1']):
-                            val = self._datacolumns[bkey].get(row)
+                            val = self._datacolumns[bkey].get(srow)
                             if val is None or abs(val - bbox[bidx]) > 2:
                                 break
                         else:
-                            return self._datacolumns[key][row]
+                            return self._datacolumns[key][srow]
             return None
 
         return annotationElementSelector
@@ -757,8 +780,7 @@ class PlottableItemData:
             if len(cols) == 4:
                 # If we load all of these from annotation elements, use all
                 # three keys:
-                # for akey in {'annotation.id', 'annotation.name', 'annotationelement.id'}:
-                for akey in {'annotationelement.id'}:
+                for akey in {'annotation.id', 'annotation.name', 'annotationelement.id'}:
                     if self._datacolumns and akey in self._datacolumns:
                         self._requiredColumns.add(akey)
                     self._ensureColumn(
@@ -991,6 +1013,10 @@ class PlottableItemData:
                                lambda record, data, row: record['label']['value'])
             self._commonColumn(columns, 'annotationelement.type', doctype, getData,
                                lambda record, data, row: record['type'])
+            self._commonColumn(columns, 'annotation.id', doctype, getData,
+                               lambda record, data, row: str(record['_aid']))
+            self._commonColumn(columns, 'annotation.name', doctype, getData,
+                               lambda record, data, row: str(record['_aname']))
             self._commonColumn(columns, 'bbox.x0', doctype, getData,
                                lambda record, data, row: record['_bbox']['lowx'])
             self._commonColumn(columns, 'bbox.y0', doctype, getData,
@@ -1015,11 +1041,13 @@ class PlottableItemData:
         for iidx, annotList in enumerate(self.annotations or []):
             iid = str(self.items[iidx]['_id'])
             for anidx, annot in enumerate(annotList):
-                # If the first item's annotation didn't contribute any required
-                # data to the data set, skip subsequent items' annotations;
-                # they are likely to be discarded.
-                if iidx and not countsPerAnnotation.get(anidx, 0) and not self._fullScan:
-                    continue
+                # This had been checking if the first item's annotation didn't
+                # contribute any required data to the data set, skip subsequent
+                # items' annotations; they are likely to be discarded.  This
+                # is untrue ui datafiles or folder level data augments the
+                # element records
+                # if iidx and not countsPerAnnotation.get(anidx, 0) and not self._fullScan:
+                #     continue
                 startcount = count
                 if annot is None:
                     continue
@@ -1029,6 +1057,8 @@ class PlottableItemData:
                 if ((not self._sources or 'annotationelement' in self._sources) and
                         Annotationelement().countElements(annot) <= self.maxAnnotationElements):
                     for element in Annotationelement().yieldElements(annot, bbox=True):
+                        element['_aid'] = annot['_id']
+                        element['_aname'] = annot['annotation']['name']
                         count += self._collectColumns(
                             columns, [element], 'annotationelement', iid=iid, aid=str(annot['_id']))
                 if not iidx:
@@ -1068,6 +1098,82 @@ class PlottableItemData:
                     countsPerDataFile[dfidx] = count - startcount
         return count
 
+    def _computeFunction(self, rows):
+        if self._compute['function'] == 'umap':
+            import umap
+
+            logger.info(f'Calling umap on {len(rows)} rows')
+            reducer = umap.UMAP(**self._compute['params'])
+            self._computed = reducer.fit_transform(list(rows.values()))
+            logger.info('Called umap')
+            return True
+
+    def _getColumnsFromCompute(self, columns):  # noqa
+        """
+        Collect columns and data from compute actions.
+        """
+
+        def computeGetData(record):
+            return {}
+
+        def computeLength(record, data):
+            return len(self._computed)
+
+        def computeSelector(key):
+            axis = ord(key[-1:]) - ord('x')
+
+            def computeSelectorAxis(record, data, row):
+                return self._computed[row][axis]
+
+            return computeSelectorAxis
+
+        if not self._datacolumns:
+            if len([col for col in columns.values() if col['type'] == 'number']) >= 2:
+                for key in self.computeColumns:
+                    title = self.commonColumns[key]
+                    self._ensureColumn(
+                        columns, key, title, 'compute', computeGetData,
+                        computeSelector(key), computeLength)
+                    columns[key]['count'] = 1
+                    columns[key]['min'] = columns[key]['max'] = 0
+            return 0
+        if self._compute is None or not len(self._requiredColumns & self.computeColumns):
+            return 0
+        compcol = {
+            key for key, col in columns.items()
+            if col['type'] == 'number' and col.get('min') is not None
+        } & set(self._compute['columns'])
+        if not len(compcol):
+            return 0
+        rows = {}
+        cols = sorted({col for col in self._compute['columns'] if col in self._datacolumns})
+        lencols = len(cols)
+        needcols = cols + sorted(set(self._requiredColumns) - set(cols) - self.computeColumns)
+        for kidx, key in enumerate(needcols):
+            for row, value in self._datacolumns[key].items():
+                if not kidx:
+                    rows[row] = [value]
+                elif row in rows and len(rows[row]) == kidx:
+                    rows[row].append(value)
+        rows = {k: row for k, row in rows.items() if len(row) == len(needcols)}
+        if not len(rows):
+            return 0
+        rows = {k: row[:lencols] for k, row in rows.items()}
+        if not self._computeFunction(rows):
+            return 0
+        for key in self.computeColumns:
+            if key in self._requiredColumns and key in self._datacolumns:
+                title = self.commonColumns[key]
+                self._ensureColumn(
+                    columns, key, title, 'compute', computeGetData,
+                    computeSelector(key), computeLength)
+                cidx = ord(key[-1:]) - ord('x')
+                for ridx, row in enumerate(rows):
+                    self._datacolumns[key][row] = float(self._computed[ridx][cidx])
+                columns[key]['count'] = len(rows)
+                columns[key]['min'] = columns[key]['max'] = 0
+        return len(rows)
+
     def _getColumns(self):
         """
         Get a sorted list of plottable columns with some metadata for each.
@@ -1086,6 +1192,7 @@ class PlottableItemData:
                     count += self._collectColumns(columns, [item], 'item', first=False)
         count += self._getColumnsFromAnnotations(columns)
         count += self._getColumnsFromDataFiles(columns)
+        count += self._getColumnsFromCompute(columns)
         for result in columns.values():
             if len(result['distinct']) <= self.maxDistinct:
                 result['distinct'] = sorted(result['distinct'])
@@ -1095,7 +1202,9 @@ class PlottableItemData:
             if result['type'] != 'number' or result['min'] is None:
                 result.pop('min', None)
                 result.pop('max', None)
-        prefixOrder = {'item': 0, 'annotation': 1, 'annotationelement': 2, 'data': 3, 'bbox': 4}
+        prefixOrder = {
+            'item': 0, 'annotation': 1, 'annotationelement': 2, 'data': 3,
+            'bbox': 4, 'compute': 5}
         columns = sorted(columns.values(), key=lambda x: (
             prefixOrder.get(x['key'].split('.', 1)[0], len(prefixOrder)), x['key']))
         return columns
@@ -1168,7 +1277,7 @@ class PlottableItemData:
             rows = [row for ridx, row in enumerate(rows) if rows[ridx] not in discard]
         return data, rows
 
-    def data(self, columns, requiredColumns=None):
+    def data(self, columns, requiredColumns=None):  # noqa
         """
         Get plottable data.
 
@@ -1182,8 +1291,14 @@ class PlottableItemData:
             columns = columns.split(',')
         if not isinstance(requiredColumns, list):
             requiredColumns = requiredColumns.split(',') if requiredColumns is not None else []
-        requiredColumns = set(requiredColumns)
+        specifiedReqColumns = set(requiredColumns)
         self._requiredColumns = set(requiredColumns)
+        if self._compute:
+            if ('compute.z' in specifiedReqColumns and
+                    self._compute['function'] == 'umap' and
+                    'n_components' not in self._compute['params']):
+                self._compute['params']['n_components'] = 3
+            self._requiredColumns.update(self._compute['columns'])
         with self._dataLock:
             self._datacolumns = {c: {} for c in columns}
             rows = set()
@@ -1201,7 +1316,7 @@ class PlottableItemData:
         for cidx, col in enumerate(colsout):
             colkey = col['key']
             numrows = len(data)
-            if colkey in requiredColumns:
+            if colkey in specifiedReqColumns:
                 data = [row for row in data if row[cidx] is not None]
             if len(data) < numrows:
                 logger.info(f'Reduced row count from {numrows} to {len(data)} '
@@ -1210,7 +1325,7 @@ class PlottableItemData:
         for cidx, col in enumerate(colsout):
             colkey = col['key']
             numrows = len(data)
-            if colkey in self._requiredColumns and colkey not in requiredColumns:
+            if colkey in self._requiredColumns and colkey not in specifiedReqColumns:
                 subdata = [row for row in subdata if row[cidx] is not None]
         if len(subdata) and len(subdata) < len(data):
             logger.info(f'Reduced row count from {len(data)} to {len(subdata)} '