+++ sources/bioformats/large_image_source_bioformats/__init__.py
@@ -23,6 +23,7 @@
 #   IFormatReader.html for interface details.
 
 import atexit
+import builtins
 import logging
 import math
 import os
@@ -105,7 +106,7 @@ def _reduceLogging():
             'org/slf4j/LoggerFactory', 'getLogger',
             '(Ljava/lang/String;)Lorg/slf4j/Logger;', rootLoggerName)
         logLevel = javabridge.get_static_field(
-            'ch/qos/logback/classic/Level', 'ERROR', 'Lch/qos/logback/classic/Level;')
+            'ch/qos/logback/classic/Level', 'OFF', 'Lch/qos/logback/classic/Level;')
         javabridge.call(rootLogger, 'setLevel', '(Lch/qos/logback/classic/Level;)V', logLevel)
     except Exception:
         pass
@@ -212,6 +213,17 @@ class BioformatsFileTileSource(FileTileSource, metaclass=LruCacheMetaclass):
         largeImagePath = str(self._getLargeImagePath())
         config._ignoreSourceNames('bioformats', largeImagePath)
 
+        header = b''
+        if os.path.isfile(largeImagePath):
+            try:
+                header = builtins.open(largeImagePath, 'rb').read(5)
+            except Exception:
+                msg = 'File cannot be opened via Bioformats'
+                raise TileSourceError(msg)
+        # Never allow pdfs; they crash the JVM
+        if header[:5] == b'%PDF-':
+            msg = 'File cannot be opened via Bioformats'
+            raise TileSourceError(msg)
         if not _startJavabridge(self.logger):
             msg = 'File cannot be opened by bioformats reader because javabridge failed to start'
             raise TileSourceError(msg)